// src/enums/customer-fields/customer-fields.ts
export enum CustomerFields {
  FIRST_NAME = 'firstName',
  LAST_NAME = 'lastName',
  EMAIL = 'email',
  DATE_OF_BIRTH = 'dateOfBirth',
  ADDRESSES = 'addresses',
  PASSWORD = 'password',
  BILLING_ADDRESS_IDS = 'billingAddressIds',
  SHIPPING_ADDRESS_IDS = 'shippingAddressIds',
  VERSION = 'version',
  DEFAULT_BILLING_ADDRESS_ID = 'defaultBillingAddressId',
  DEFAULT_SHIPPING_ADDRESS_ID = 'defaultShippingAddressId',
}